Suggestion: remove `fiber.new`
The `fiber.new` functionality can be implemented in terms of executing `fiber.spawn` with a function that immediately suspends and switches back to the parent, passing its own identity as a payload. `fiber.spawn` can also be implemented in terms of `fiber.new`, but `fiber.new` is less optimizable in the case of spawning a fiber that returns normally without ever suspending and also requires an awkward new function syntax for essentially having multiple function entry points. Overall it seems like a nice simplification with minimal downside to keep only `fiber.spawn`. Is there some benefit of `fiber.new` that I'm missing?

The only benefit I can think of would be potential optimizations to lazily allocate resources to help the case where a fiber is created (in a paused state) but then never entered. But the toolchain can do the same optimizations by calling `fiber.spawn` only once it knows the fiber will actually be entered.