Suggestion: consolidate fiber switching mechanisms
We can consolidate suspending and resuming into a single mechanism by 1) making the simplifying assumption that the toolchain can always arrange for suspending fibers have access to the identity of the next fiber that should run, and 2) considering resume ancestors to be suspended and eligible to switch to.

In particular, `fiber.suspend f` is analogous to `fiber.resume_to parent(f)`, `fiber.resume f` is analogous to `fiber.resume_to f`.

This gives a flatter and more uniform model of fibers in which the resume stack controls normal and exceptional returns from fibers, but has no bearing on how fibers otherwise switch between each other.

Note that with this change the root fiber needs an identity so it can be switched to. I don't see any downside to having a `fiber.current` instruction to facilitate this. That would also allow the fiber-self argument to fiber functions to be removed.

Note that a further simplification would be to remove `fiber.switch`, since AFAICT toolchains should be able to have the resume ancestor participate in the switch. I've kept that instruction in for now to minimize the delta.

We could also remove `fiber.retire` (in favor of `fiber.retireto`) under the assumption that the toolchain can make switch targets available. And we could remove `fiber.retireto` just like we could remove `fiber.switch` under the assumption that the toolchain can arrange for resume ancestors to participate in fiber management.