Refactor `cont.wast` to not use `let` & `func.bind` (#24)
This commit refactors the file `cont.wast` such that it no longer uses
`let` and `func.bind`. We change the type of `spawn` such that it is
parameterised by a continuation type rather than a function type, and
then we replace every use of `func.bind` by a corresponding
`cont.bind`.