Update the instruction set descriptions in the Explainer document (#27)
This patch brings the description of the proposed instruction set extension in sync with the reference implementation. It also fixes a few other bugs such as a dead link in the TOC and the syntax in the examples.

The document still warrants a refactoring as some things could do with better explanations, some TODOs left to complete, and some of the examples still make use of deprecated features from the function-references proposal.