[BoS] Rename instructions
`switch` becomes `stack.switch` because we are using the `stack` prefix for all
bag-o-stacks instructions.

`switch.call` becomes `stack.new_switch` because it is the composition of
`stack.new` (not yet documented) and `stack.switch`.

Similarly, `switch.return` becomes `stack.switch_retire` because it performs a
switch and retires the previous active stack. We use "retire" instead of
"return" because "return" is not really an accurate description of what is
happening.