[BoS] Rename instructions (#51)
`switch.call` becomes `stack.new_switch` because it is the composition of
`stack.new` (not yet documented) and `switch`.

Similarly, `switch.return` becomes `switch_retire` because it performs a switch
and retires the previous active stack. We use "retire" instead of "return"
because "return" is not really an accurate description of what is happening.