[BoS] `stack.new` takes a suffix rather than prefix of func args
The `stack.new` instruction binds some argument to the function that will be
called on the new stack, leaving the rest to be supplied by `stack.switch`.
Previously the arguments bound by `stack.new` were a prefix of the function
arguments, but change it so that they are a suffix of the function
arguments (besides the return stack reference) instead.

This allows us to relax the typing of `stack.new_switch` to be able to use stack
types that only send a prefix of the function arguments while maintaining the
property that `(stack.new_switch x y)` can be rewritten as `(stack.new x
y) (stack.switch x)`. It also gives us the new property that `(stack.new x
y) (stack.switch x)` can be rewritten as `(stack.new_switch x y)`; previously
that was only true if the parameters of the stack type at `x` matched the
parameters of the function `y` or if the stack type had no parameters besides
the return stack reference.