Add full code for $scheduler1 and $scheduler2 examples (#83)
The third section of the explainer currently compares task scheduling using `resume`/`suspend` vs `switch` using two example modules, `$scheduler1`  and `$scheduler2`. The explainer only contains a skeleton of the actual code.

This PR adds `.wast` files containing full code for these modules and links to them in the explainer. The code is carefully written to follow the structure shown in the explainer.