Fix validation of `switch`. (#99)
* Fix validation of `switch`.

The validation of `switch` was not working as intended when the
"switcher" and "switchee" had different continuation type immediates.

The fix is to replace the current continuation from the argument list
with the switched-to continuation. The previous thing happened to be
working when the "switcher" and "switchee" used the same continuation
type immediate.

I also noticed a bug in the testsuite runner. It was not running the stack switching tests since commit https://github.com/WebAssembly/stack-switching/commit/70086b9093716c2ffc958eff06a8a3bb73b9a8a2. I have added the "stack-switching" subdirectory to the test script runner such that the tests are now being run again by `make test`.

Resolves #98.

* Update interpreter/valid/valid.ml

Co-authored-by: Andreas Rossberg <rossberg@mpi-sws.org>

---------

Co-authored-by: Andreas Rossberg <rossberg@mpi-sws.org>