Reduction semantics (#109)
* Reduction semantics

This patch populates the "Execution" section of the Explainer document
with the reduction rules for stack switching.

* Some type expansions

* Fix minor typo, hdl => hdl1

* Fix typo: stray '

* Address Andreas' feedback

* Remove stray ^

* Incorporated comments

Incorporated comments from before
The biggest thing I am unsure of is lines 814 and 826. This pertains to what point the tag argument of suspend and switch is translated into a tag address. Doing it in the reduction rule for resume (as was the case in the previous version of Explainer.md) is wrong, for reasons explained in the comment I wrote at the time. I believe I was told in the actual implementation, the change happens during the validation phase, which is why I propose these lines 814 and 826; however my phrasing can perhaps be improved.

* rectified the suspend address translation to its correct place

Incorporating Thomas Lively's comment, I have rectified the spot where suspend and switch translate the tag index into a tag address, by creating new instructions suspend.addr and switch.addr that take addresses (rather than indices) as arguments. This allows for a rule focused on translating suspend to suspend.addr (and likewise for switch), enforcing that the correct wasm frame F is used.

* Update proposals/stack-switching/Explainer.md

Co-authored-by: Thomas Lively <tlively123@gmail.com>

* Fix resume_throw semantics

* Update proposals/stack-switching/Explainer.md

Committing tlively's suggested change

Co-authored-by: Thomas Lively <tlively123@gmail.com>

* Update proposals/stack-switching/Explainer.md

Committing tlively's suggested change

Co-authored-by: Thomas Lively <tlively123@gmail.com>

* Applied all suggested changes except taguse

Applied all suggested changes except taguse which is the last remaining issue

* Added tag uses

* using taguse in hdl too

* factored typing rules and added keyword 'on' where now necessary

* Typo in typing rule for tag addresses

* corrected $a to $e and added switch failure rules

---------

Co-authored-by: mlegoupil <90708290+mlegoupil@users.noreply.github.com>
Co-authored-by: Sam Lindley <Sam.Lindley@ed.ac.uk>
Co-authored-by: Thomas Lively <tlively123@gmail.com>